# bpo-46055: Streamline inner loop for right shifts