While reviewing #30044, I noticed that the inner loop for the right shift operation could be more efficient. Here's a PR that streamlines that loop. The main changes are:

- remove an unnecessary extra masking operation (`& lomask`), and replace `& himask` with `& PyLong_MASK`
- rewrite the loop to remove multiple accesses to the digits of `a` and `z`
- remove an inner branch

On my machine (macOS 10.14.6 / Intel MacBook Pro), in informal timings I get approximately a 35% speedup for a shift of the form `huge >> small`. Some sample timings:

On main (commit cf15419a99962b3503ff7123e148d71f7f01bc15):
```
lovelace:cpython mdickinson$ ./python.exe -m timeit -s "a, b = 7**100000, 53" "a >> b"
20000 loops, best of 5: 11.5 usec per loop
```

On this branch (commit 056495d5cde239ab19920a11c07400bdcfe1ec85):
```
lovelace:cpython mdickinson$ ./python.exe -m timeit -s "a, b = 7**100000, 53" "a >> b"
50000 loops, best of 5: 8.51 usec per loop
```

Small shift operations are not significantly affected. More sample timings - on master:

```
lovelace:cpython mdickinson$ ./python.exe -m timeit -s "a, b = 7**10, 53" "a >> b"
10000000 loops, best of 5: 23.3 nsec per loop
```

On this branch:

```
lovelace:cpython mdickinson$ ./python.exe -m timeit -s "a, b = 7**10, 53" "a >> b"
10000000 loops, best of 5: 24 nsec per loop
```

(but a second run on this branch gave 22.9 nsec per loop, so any difference is being lost in the variation between runs)

<!-- issue-number: [bpo-46055](https://bugs.python.org/issue46055) -->
https://bugs.python.org/issue46055
<!-- /issue-number -->