@mdickinson 2021-12-23
New timings with the latest commit (f92f3f80741df576012fa01c19d341e0e00a65c9) give me a ~75% speedup for the `big >> small` case.

On this branch:
```
lovelace:cpython mdickinson$ git checkout main && make >> /dev/null && ./python.exe -m timeit -s "a, b = 7**100000, 53" "a >> b"
[ ... build output snipped ...]
20000 loops, best of 5: 11.6 usec per loop
lovelace:cpython mdickinson$ git checkout faster-rshift && make >> /dev/null && ./python.exe -m timeit -s "a, b = 7**100000, 53" "a >> b"
[ ... build output snipped ...]
50000 loops, best of 5: 6.53 usec per loop
```