@mdickinson 2021-12-27
> I am wondering how it works on 32-bit platform with 32-bit digit

Yes, me too. I'll see if I can find a way to check this.