@mdickinson 2021-12-27
> I'll run some timings and look at the generated code. If the cast-first variant is faster, I'll change this.

At least on my machine, I'm not seeing any difference. The assembly generated for the inner loop is identical both ways (`z->ob_digit[i] = (digit)accum & PyLong_MASK;` versus `z->ob_digit[i] = (digit)(accum & PyLong_MASK);`, except for a difference in column numbers in the `.loc` statements.

```
Ltmp3675:
LBB51_18:
	##DEBUG_VALUE: long_lshift1:remshift <- [DW_OP_constu 44, DW_OP_minus] [$rbp+0]
	##DEBUG_VALUE: long_lshift1:j <- $rsi
	##DEBUG_VALUE: long_lshift1:accum <- $rax
	##DEBUG_VALUE: long_lshift1:z <- $r12
	##DEBUG_VALUE: long_lshift1:i <- $r15
	##DEBUG_VALUE: long_lshift1:a <- $rbx
	##DEBUG_VALUE: long_lshift1:newsize <- $r13
	.loc	1 4589 29 is_stmt 1     ## Objects/longobject.c:4589:29
	movl	24(%rbx,%rsi,4), %edx
                                        ## kill: def $cl killed $cl killed $rcx
	.loc	1 4589 44 is_stmt 0     ## Objects/longobject.c:4589:44
	shlq	%cl, %rdx
	.loc	1 4589 15               ## Objects/longobject.c:4589:15
	orq	%rax, %rdx
Ltmp3676:
	##DEBUG_VALUE: long_lshift1:accum <- $rdx
	.loc	1 4590 26 is_stmt 1     ## Objects/longobject.c:4590:26
	movl	%edx, %eax
	andl	$1073741823, %eax       ## imm = 0x3FFFFFFF
	.loc	1 4590 24 is_stmt 0     ## Objects/longobject.c:4590:24
	movl	%eax, 24(%r12,%r15,4)
	.loc	1 4591 15 is_stmt 1     ## Objects/longobject.c:4591:15
	shrq	$30, %rdx
Ltmp3677:
	##DEBUG_VALUE: long_lshift1:accum <- $rdx
	movq	%rdx, %rax
```

And here's the corresponding code with the code as it currently is (mask, then cast)