@mdickinson 2021-12-27
Whoops; sorry - that assembly excerpt was from changing the `long_lshift1` inner loop, not the `long_rshift1` one. But I see the same with `long_rshift1`: no non-trivial change to the assembly code. (The details are a bit messier: it looks as though clang has decided to unroll the inner loop for `long_rshift1`.)